Pages, Layout and styles
In order to have the navigation working I needed to remove all pages and add them again. Afterward custom @imports and other functionality was throwing errors so they were removed and navigation worked. Compare these pages to the one in the master to see changes and re-add hat is needed.